Start from scratch with Onsen
Use the Onsen framework to ensure a future-proof code base. This
starting point is more or less a clone of https://github.com/frandiox/OnsenUI-Todo-App, and therefore is redistributed with the MIT OpenSource license. Open source fighting!

Note that this setup, even though it only uses HTML, CSS and JavaScript,
requires you to not just open the `index.html` file, but to host it with
e.g. Web Server For Chrome (https://chrome.google.com/webstore/detail/web-server-for-chrome/ofhbbkphhbklhfoeikjpcbhemlocgigb)